+++ packages/otel/src/index.ts
@@ -1,2 +1,2 @@
 export { GenAIOpenTelemetryIntegration } from './gen-ai-open-telemetry-integration';
-export { OpenTelemetry } from './open-telemetry-integration';
+export { OpenTelemetry } from './open-telemetry';
